+++ posthog/src/main/java/com/posthog/internal/PostHogDecideRequest.kt
@@ -6,14 +6,16 @@ package com.posthog.internal
 internal class PostHogDecideRequest(
     apiKey: String,
     distinctId: String,
-    anonymousId: String,
+    anonymousId: String?,
     groups: Map<String, Any>?,
     // add person_properties, group_properties
 ) : HashMap<String, Any>() {
     init {
         this["api_key"] = apiKey
         this["distinct_id"] = distinctId
-        this["\$anon_distinct_id"] = anonymousId
+        if (!anonymousId.isNullOrBlank()) {
+            this["\$anon_distinct_id"] = anonymousId
+        }
         if (groups?.isNotEmpty() == true) {
             this["\$groups"] = groups
         }